Issue-184 - Refactored visualizer util classes so they can be included in test coverage
