+++ client/src/main/java/com/github/goldy1992/mp3player/client/models/visualizers/fountainspring/Constants.kt
@@ -1,0 +1,5 @@
+package com.github.goldy1992.mp3player.client.models.visualizers.fountainspring
+
+object Constants {
+    const val g = -9.8f
+}
