+++ client/src/main/java/com/github/goldy1992/mp3player/client/ui/components/equalizer/bar/BarEqualizer.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
@@ -17,7 +16,7 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.goldy1992.mp3player.client.ui.DpPxSize
-import com.github.goldy1992.mp3player.client.utils.calculateBarSpacingPixels
+import com.github.goldy1992.mp3player.client.utils.visualizer.calculateBarSpacingPixels
 
 private const val MAX_AMPLITUDE = 400f
 private const val LOG_TAG = "BarEqualizer"
