+++ client/src/main/java/com/github/goldy1992/mp3player/client/ui/components/equalizer/cards/FountainSpringCard.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.github.goldy1992.mp3player.client.R
-import com.github.goldy1992.mp3player.client.ui.components.equalizer.fountainspring.FountainSpringEqualizer
+import com.github.goldy1992.mp3player.client.ui.components.equalizer.fountainspring.FountainSpringVisualizer
 
 @Preview
 @Composable
@@ -18,7 +18,7 @@ fun FountainSpringCard(
         modifier = modifier,
         title = stringResource(id = R.string.fountain),
     ) { canvasSize, containerModifier ->
-        FountainSpringEqualizer(
+        FountainSpringVisualizer(
             frequencyPhasesProvider = frequencyPhases,
             canvasSize = canvasSize,
             isPlayingProvider = isPlaying,
