--- client/src/main/java/com/github/goldy1992/mp3player/client/ui/components/equalizer/fountainspring/Particle.kt
@@ -1,29 +1,0 @@
-package com.github.goldy1992.mp3player.client.ui.components.equalizer.fountainspring
-
-
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
-
-private const val logTag = "Particle"
-
-data class Particle
-constructor(
-    val startPosition : Offset,
-    val removePoint : Float = startPosition.y * 0.9f,
-    val hMax : Float,
-    val isFalling : Boolean,
-    val initialVelocity : Float = 7.5f,
-    val angle : Double = 90.0,
-    val radius : Float = 50f,
-    var width : Float = 10f,
-    val color : Color = Color.Blue,
-    val x : Float = 0f,
-    val y : Float = 0f,
-    val elapsedTimeDeciSeconds : Float = 0f,
-) {
-
-    override fun toString(): String {
-        return "x: $x, y: $y, elapsedTimeSecs: $elapsedTimeDeciSeconds, startPos: $startPosition, initialVelocity: $initialVelocity, angle $angle"
-    }
-}
-
