+++ client/src/main/java/com/github/goldy1992/mp3player/client/utils/TimeUtils.kt
@@ -3,11 +3,12 @@ package com.github.goldy1992.mp3player.client.utils
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
+import kotlin.math.pow
 
 /**
  * Util Class for the timer
  */
-object TimerUtils {
+object TimeUtils {
     /**
      * One second constant
      */
@@ -24,5 +25,13 @@ object TimerUtils {
         val timerFormat = SimpleDateFormat("mm:ss", Locale.getDefault())
         //Log.d(LOG_TAG, "returning formatted time: " + formattedTime);
         return timerFormat.format(date)
+    }
+
+    /**
+     * From nano seconds (i.e. 10^-9)
+     * Runtime speed is Deci-seconds (i.e. 10^-1)
+     */
+    fun convertNanoSecondsToRuntimeSpeed(valueNs: Long) : Float {
+        return valueNs * 10.0.pow(-8.0).toFloat()
     }
 }
